Replace uint32 with uint32_t.
First, this is in line with NPString, next, this is in line with the rest of the file.

Finally, I was lucky to have a toolchain where npapi.h does not define uint32 at all, thus toolchain failed to compile this file. I still can't figure out how that different npapi.h can appear, but even unrelated to this the fix seems useful.

Review URL: http://codereview.chromium.org/2843045

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@457 050acbb0-2703-11df-ab0a-9f3f633ae91d