Remove CygWin form the SDK installer.
In this CL: Vast, sweeping hacks to dig CygWin out of the SDK installer

NOT in this CL: total removal of CygWin.  This CL uses the old bash-script
method of building the .nsi files, complete with the patch.  Note that the
.nsi file has shrank from > 20k lines to about 2k lines.  The patch file is
also free of CygWin stuff.  The build bot still needs CygWin in order to
build the installer, but once the installer is made, the SDK no longer
requires CygWin.

Also NOT in this CL: installation of python.  This version of the SDK
installer assumes you will install python separately.  I'd like to work
through the review of this CL before adding python to it.  Mostly because
I would like to get rid of CygWin entirely - even for the installer build,
and it might make sense to do all that at one time.
Review URL: http://codereview.chromium.org/7014022

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@853 050acbb0-2703-11df-ab0a-9f3f633ae91d