Remove build_tools/download_compilers.py and use download_toolchain.py instead
Rolled NaCl DEPS to 6511 and Toolchain DEPS to 6494.

This change puts the SDK and NaCl repositories in better sync by using more of the same tools.  It also makes it easier to add more toolchains to the SDK in the future.

Note that after this change, the toolchain directory will be renamed to match that of NaCl; specifically, the name will have _newlib appended to the end.  Example:

toolchain/linux_x86 -> toolchain/linux_x86_newlib

BUG=None
TEST=installer_test

Review URL: http://codereview.chromium.org/7711017

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1062 050acbb0-2703-11df-ab0a-9f3f633ae91d