Move tumbler into the examples directory.
Note: this requires build 97490 or later.  Canary on Mac is 97825.

There are problems with the 3D - it doesn't flush every frame for some reason, and I don't really like everything about opengl_context, but I wanted to get this moved out of experimental so we can start working on it for real.
Review URL: http://codereview.chromium.org/7701016

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1069 050acbb0-2703-11df-ab0a-9f3f633ae91d