Fix the table of contents function to return *all* the necessary info.
The function that gets the table of contents from a tar archive was incomplete.
It only returned lists of files, symlinks, etc. - but in the case of linked
files it failed to return the source of the link.  This change makes the
function return files and dirs as sets (so they are quickly searchable) and
symlinks and hlinks as dictionaries, whose values are the source of the links.

I added some more symlinks and hlinks in the test tarball to test links that
are in other dirs from their source files.
Review URL: http://codereview.chromium.org/7795029

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1116 050acbb0-2703-11df-ab0a-9f3f633ae91d