Make the merge operator in PathSet smarter.
When merging PathSets, any keys that overlap between (symlinks, links) and (files, dirs) need to be discarded from (files, dirs).

This means that when merging a TarArchive created by walking a directory tree with a TarArchive created from a
manifest file will produce the expected results.  That is, any files or dirs discovered when walking the directory
tree will be correctly replaced with the corresponding links, if those links are enumerated in the manifest.

BUG=none
TEST=run_path_set_test
Review URL: http://codereview.chromium.org/7922016

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1161 050acbb0-2703-11df-ab0a-9f3f633ae91d