Use tarfile to extract installer in sdk_update.
Change the extractor to use tarfile instead of calling tar directly.  This
allows using tar archives on Windows.  Changed the extractor logic to use NSIS
extraction if the installer archive ends with '.exe', otherwise tar.

Also in this change: the Mac & Linux tarballs no longer contain the
native_client_sdk_x_x_x prefix.  This is so the --strip-component option to tar
is not required.  This option is difficult to support when using python's
tarfile module without doing file system moves.

Also in this change: add pong to the whitelist of examples.
Review URL: http://codereview.chromium.org/7982031

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1191 050acbb0-2703-11df-ab0a-9f3f633ae91d