I found a bug.  I had fixed a 'lockup' bug earlier in the week, but hit a new one while testing yesterday.  This should be the last bug. ;-)
At startup, if the user clicks before the nexe is really running
(even though the module has barely loaded), then the first PostMessage (of the
user's first move) is just never seen (HandleMessage does not fire).

But, the JS user interface has sent the message and thinks it the user's turn...
but the computer is waiting for the user's move...

The solution is to have the UI wait for the AI to send a 'ready' message before
allowing any moves.
Review URL: http://codereview.chromium.org/8561013

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1378 050acbb0-2703-11df-ab0a-9f3f633ae91d