Split NaCl VS platform into NaCl64 and NaCl32
Most of the files now get installed to MSBuild folder
but not under "Platforms".  The platform specific stuff is
now in NaCl64, which gets duplicated  (with replacements)
to NaCl32 during create_package.

The result is two platforms which are almost identical
except for TargetArchitecture (which is no longer settable
in the GUI).

BUG=151687

Review URL: https://codereview.chromium.org/11091007

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1436 050acbb0-2703-11df-ab0a-9f3f633ae91d