[VS Addin] Don't fail installation if win32 platform is missing.
It seems that it is possible to install VS with only
x86 and not win32 platforms.  In this case we can't
install the PPAPI platform (as it is a clone of win32)
but we should probably not fail the installation
completely.

Also, handle errors from create_ppapi_platform more gracefully.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/19990004

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1564 050acbb0-2703-11df-ab0a-9f3f633ae91d