[VS AddIn] Fix default library paths.
Also, add pnacl-finalize step to PNaCl build.

The default library path for all platforms was wrong and
did not include the location of the SDK libraries such
as ppapi_cpp.  There was a bug in the test code that meant
that we didn't catch this since we where not actually linking
any of the test projects to .nexe or .pexe files.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/69343013

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1567 050acbb0-2703-11df-ab0a-9f3f633ae91d