Fix buildbot upload of vs_addin archive.
This change adddresses a couple of issues with the upload
of the vs_addin archive to google storage.

Firstly is disables the upload step for periodic builders
so that they don't clobber existing uploads.

Secondly we check for the existence of a file on google
storage before uploading and error out if one is already
found.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/124123002

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1569 050acbb0-2703-11df-ab0a-9f3f633ae91d