[VS Addin] Fix several issues with install script.
Fix the --vsuser-path flag.  It was attempted to append
to a tuble.

Don't depend on any one of the vsuser-paths existing, only
that at least one of them exists.  This allows installation
to succeed on platforms that only have one version of VS
installed.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/282153003

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1573 050acbb0-2703-11df-ab0a-9f3f633ae91d